+++ src/control_plane/workload/mod.rs
@@ -3,8 +3,12 @@ mod tests;
 
 mod application_container_plan;
 mod application_container_plan_options;
+mod application_container_request;
+mod application_container_request_options;
 mod workload_plan_error;
 
 pub(crate) use application_container_plan::ApplicationContainerPlan;
 pub(crate) use application_container_plan_options::ApplicationContainerPlanOptions;
+pub(crate) use application_container_request::application_container_request;
+pub(crate) use application_container_request_options::ApplicationContainerRequestOptions;
 pub(crate) use workload_plan_error::WorkloadPlanError;